+++ docs/app.js
@@ -1,8 +1,0 @@
-import Vue from 'vue';
-import { VueCsvImport } from '../src';
-
-new Vue({
-    components: { VueCsvImport },
-
-    el: '#app',
-});